Adds support for scheduled events service (#8)
Adds a concept of a scheduled events service. This service is able to take any arbitrary object and a time it wants to be reminded at. Then when that time happens, it triggers a callback.
The events are all persisted in the selected storage
Now in memory_module, the episodic memory extraction is also triggered when a certain amount of time passes.
I added tests to test it all out
Also added an e2e test in test_memory_module to test the functionality out