Add LLM Config to Config and remove LLMService requirement for memory module (#10)
LLM Service used to be a requirement for MemoryModule. This meant the client needed to build this and pass it in. In this PR, we include all its arguments as part of the MemoryModuleConfig.
This means LLMService needs not be a required service to the module. But the config is.