Fix timezone issue for sqlite + only check for single author for user messages (#71)
- Sqlite has an annoying issue where datetimes are not automatically
stored as utc. So I have to do it manually. It uses local time by
default. Grr.
- It has another issue where when comparing dates, if your date doesn't
have a T in it (like isoFormat), it doesn't know how to compare it
correctly with datetime('now'). So you need to format it correctly using
`strftime`. It's really, the worst.
- Finally, we had a check during extraction that would force us to make
sure the authors in a conversation were unique. We used to save
author_id as user_id for assistant messages. But now we store bot_id. So
this fixes that. We now check if there is a unique author in only the
"user" messages.