memory module project & sample changes (#112)
This PR introduces some changes that I had to make to be able to deploy
the tech assistant agent to Azure, and also some other things.

1. Restructured the `packages/memory_module` folder by putting all the
code files in the `packages/memory_module/memory_module` folder. This
project setup is needed to correctly build the memory_module into a
distribution file that can be pip-installed.
* Updated references in all other packages, the sample, tests as well as
vscode settings.
1. Renamed the memory module project name from `memory-module` to
`memory_module` and set the version to `0.1.0`.
1. Addded docs and guidelines for how to test the sample in Test Tool,
Teams, and also deploy it to Azure. The flow to deploy Azure is a little
bit clunky the memory_module package has to be zipped and added to the
`src/dist` folder. Also there's no clean way to support both Azure
OpenAI or OpenAI configs since TTK complains about missing fields when
mapping env variables to the Azure bicep templates.
1. Fixed a bunch of random ruff and mypy linting errors.