Add scoped memory module (#109)
1. This PR fixes some stuff that was preventing the bot from working
(these are minor changes)
2. Added scoped memory module. The main reason for this is that certain
functions in memory module required things like "conversation_ref" or
"user_id" which are tedious to extract from context and pass in
everywhere. And that make for poorer dev-ex. This scoped memory module
scopes the memory module to a particular conversation / set of users.
3. This scoped memory module now also augments the context, and now is
available as part of the turncontext via `context.get('memory_module')`
4. This also made it such that the MemoryMiddleware didn't actually need
a memory_module parameter. It could just take config args and build
memory_module directly for the user, which further simplified
construction. (1 thing as opposed to 2)