Add a storage config (#121)
We have an llmConfig, but for storage we kept all the details in a
top-level config (db_path). In this PR, we follow the same path for
storage configs. These storage configs can have arbitrary values, and
can be passed into future Storage types as well.

Why:
I was writing documentation for it, and the config for storage didn't
follow the same pattern as the config for the llm which was weird. This
fixes that.