Improve llm_service completions typings (#143)
# Summary of PR

In this PR, we update the way we create our client to be more reflective
of the way that instructor wants us to do it
(https://docs.litellm.ai/docs/tutorials/instructor). This removes our
need for type ignores.

I also added better overrides so that if you call the completion
function without `response_model` it would return `ModelResponse`,
otherwise it would return `T`.

# Testing

Tested manually, and unit tests

# Checklist

- [x] I thought through documentation (and updated the docs if
necessary)
- [x] I updated the CHANGELOG.md file (or have a good reason not to)