Remove requirement to supply topic or query to search_memories (#147)
# Summary of PR

Previously, topic or query was a requirement for `search_memories`. But
if the dev wants to look at all memories for a user, then this function
shouldn't assert this requirement. They may have a good reason to do
this (Maybe they want to answer a question using all the previous
knowledge of the user)

# Details

<!---
Not needed if it's trivial or the summary covers it
-->

# Testing

Added a unit test

# Checklist

- [x] I thought through documentation (and updated the docs if
necessary)
- [x] I updated the CHANGELOG.md file (or have a good reason not to)