Fix `configure_logging` bug (#151)
# Summary of PR
If `configure_logging` is called twice, then two handlers will be added
to the "teams_memory" logger. And so a single log statement (ex
`logger.info("hi")` will be printed to the console twice.

This PR fixes that issue by checking if the "teams-memory-console"
handler was already added to the logger.

# Checklist

- [x] I thought through documentation (and updated the docs if
necessary)
- [x] I updated the CHANGELOG.md file (or have a good reason not to)